Opengl ES 2.0 start load blender OBJ file / draw rectangle / draw text / lighted texture shaders
This is a starting point for Android Studio Opengl ES 2.0. This loads a OBJ file (made with Blender), single texture only. Each OBJ vertex can only have one UV texture coordinate. So, you might have to increase the number of verticies for complicated texture wrapping. The OBJ model can be drawn with transparency, lighted, or with alpha values. It also renders a 4 vertex rectangle. And, draws text to a rectangle.